скрипт к спелу 147655 - Токсин хваткой лягушки возможно переполнение типа

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -952,7 +952,7 @@ void ObjectMgr::LoadEquipmentTemplates()
     {
         Field* fields = result->Fetch();
 
-        uint16 entry = fields[0].GetUInt16();
+        uint32 entry = fields[0].GetUInt32();
 
         EquipmentInfo& equipmentInfo = _equipmentInfoStore[entry];
